+++ angular/src/directives/navigation/ion-tabs.ts
@@ -1,6 +1,5 @@
-import { Component, ContentChild, EventEmitter, HostListener, Input, Output, ViewChild } from '@angular/core';
+import { Component, ContentChild, EventEmitter, HostListener, Output, ViewChild } from '@angular/core';
 
-import { EnvironmentInjector } from '../../di/r3_injector';
 import { NavController } from '../../providers/nav-controller';
 import { IonTabBar } from '../proxies';
 
@@ -11,12 +10,7 @@ import { StackEvent } from './stack-utils';
   selector: 'ion-tabs',
   template: ` <ng-content select="[slot=top]"></ng-content>
     <div class="tabs-inner">
-      <ion-router-outlet
-        #outlet
-        tabs="true"
-        [environmentInjector]="environmentInjector"
-        (stackEvents)="onPageSelected($event)"
-      ></ion-router-outlet>
+      <ion-router-outlet #outlet tabs="true" (stackEvents)="onPageSelected($event)"></ion-router-outlet>
     </div>
     <ng-content></ng-content>`,
   styles: [
@@ -51,16 +45,6 @@ import { StackEvent } from './stack-utils';
 export class IonTabs {
   @ViewChild('outlet', { read: IonRouterOutlet, static: false }) outlet: IonRouterOutlet;
   @ContentChild(IonTabBar, { static: false }) tabBar: IonTabBar | undefined;
-
-  /**
-   * @experimental
-   *
-   * The `EnvironmentInjector` provider instance from the parent component.
-   * Required for using standalone components with `ion-router-outlet`.
-   *
-   * Will be deprecated and removed when Angular 13 support is dropped.
-   */
-  @Input() environmentInjector: EnvironmentInjector;
 
   @Output() ionTabsWillChange = new EventEmitter<{ tab: string }>();
   @Output() ionTabsDidChange = new EventEmitter<{ tab: string }>();