+++ angular/test/base/src/app/overlays-inline/overlays-inline-routing.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from "@angular/core";
+import { RouterModule } from "@angular/router";
+import { OverlaysInlineComponent } from "./overlays-inline.component";
+
+@NgModule({
+  imports: [RouterModule.forChild([
+    {
+      path: '',
+      component: OverlaysInlineComponent
+    }
+  ])],
+  exports: [RouterModule]
+})
+export class OverlaysInlineRoutingModule { }