+++ core/src/components/item-options/item-options.tsx
@@ -2,8 +2,8 @@ import type { ComponentInterface, EventEmitter } from '@stencil/core';
 import { Component, Element, Event, Host, Method, Prop, h } from '@stencil/core';
 
 import { getIonMode } from '../../global/ionic-global';
-import type { Side } from '../../interface';
 import { isEndSide } from '../../utils/helpers';
+import type { Side } from '../menu/menu-interface';
 
 @Component({
   tag: 'ion-item-options',