+++ core/src/components/radio/radio.tsx
@@ -3,11 +3,17 @@ import { Component, Element, Event, Host, Method, Prop, State, Watch, h } from '
 
 import { getIonMode } from '../../global/ionic-global';
 import type { Color, StyleEventDetail } from '../../interface';
-import { addEventListener, getAriaLabel, removeEventListener } from '../../utils/helpers';
+import type { LegacyFormController } from '../../utils/forms';
+import { createLegacyFormController } from '../../utils/forms';
+import type { Attributes } from '../../utils/helpers';
+import { addEventListener, getAriaLabel, inheritAriaAttributes, removeEventListener } from '../../utils/helpers';
+import { printIonWarning } from '../../utils/logging';
 import { createColorClasses, hostContext } from '../../utils/theme';
 
 /**
  * @virtualProp {"ios" | "md"} mode - The mode determines which platform styles to use.
+ *
+ * @slot - The label text to associate with the radio. Use the "labelPlacement" property to control where the label is placed relative to the radio.
  *
  * @part container - The container for the radio mark.
  * @part mark - The checkmark or dot used to indicate the checked state.
@@ -24,6 +30,11 @@ export class Radio implements ComponentInterface {
   private inputId = `ion-rb-${radioButtonIds++}`;
   private radioGroup: HTMLIonRadioGroupElement | null = null;
   private nativeInput!: HTMLInputElement;
+  private legacyFormController!: LegacyFormController;
+  private inheritedAttributes: Attributes = {};
+
+  // This flag ensures we log the deprecation warning at most once.
+  private hasLoggedDeprecationWarning = false;
 
   @Element() el!: HTMLIonRadioElement;
 
@@ -71,6 +82,37 @@ export class Radio implements ComponentInterface {
   }
 
   /**
+   * Where to place the label relative to the radio.
+   * `"start"`: The label will appear to the left of the radio in LTR and to the right in RTL.
+   * `"end"`: The label will appear to the right of the radio in LTR and to the left in RTL.
+   * `"fixed"`: The label has the same behavior as `"start"` except it also has a fixed width. Long text will be truncated with ellipses ("...").
+   */
+  @Prop() labelPlacement: 'start' | 'end' | 'fixed' = 'start';
+
+  // TODO FW-3125: Remove the legacy property and implementation
+  /**
+   * Set the `legacy` property to `true` to forcibly use the legacy form control markup.
+   * Ionic will only opt components in to the modern form markup when they are
+   * using either the `aria-label` attribute or the default slot that contains
+   * the label text. As a result, the `legacy` property should only be used as
+   * an escape hatch when you want to avoid this automatic opt-in behavior.
+   * Note that this property will be removed in an upcoming major release
+   * of Ionic, and all form components will be opted-in to using the modern form markup.
+   */
+  @Prop() legacy?: boolean;
+
+  /**
+   * How to pack the label and radio within a line.
+   * `"start"`: The label and radio will appear on the left in LTR and
+   * on the right in RTL.
+   * `"end"`: The label and radio will appear on the right in LTR and
+   * on the left in RTL.
+   * `"space-between"`: The label and radio will appear on opposite
+   * ends of the line with space between the two elements.
+   */
+  @Prop() justify: 'start' | 'end' | 'space-between' = 'space-between';
+
+  /**
    * Emitted when the styles change.
    * @internal
    */
@@ -103,36 +145,52 @@ export class Radio implements ComponentInterface {
   }
 
   connectedCallback() {
+    this.legacyFormController = createLegacyFormController(this.el);
     if (this.value === undefined) {
       this.value = this.inputId;
     }
     const radioGroup = (this.radioGroup = this.el.closest('ion-radio-group'));
     if (radioGroup) {
       this.updateState();
-      addEventListener(radioGroup, 'ionChange', this.updateState);
+      addEventListener(radioGroup, 'ionValueChange', this.updateState);
     }
   }
 
   disconnectedCallback() {
     const radioGroup = this.radioGroup;
     if (radioGroup) {
-      removeEventListener(radioGroup, 'ionChange', this.updateState);
+      removeEventListener(radioGroup, 'ionValueChange', this.updateState);
       this.radioGroup = null;
     }
   }
 
   componentWillLoad() {
     this.emitStyle();
+
+    if (!this.legacyFormController.hasLegacyControl()) {
+      this.inheritedAttributes = {
+        ...inheritAriaAttributes(this.el),
+      };
+    }
   }
 
-  @Watch('color')
   @Watch('checked')
+  @Watch('color')
   @Watch('disabled')
-  emitStyle() {
-    this.ionStyle.emit({
-      'radio-checked': this.checked,
+  protected styleChanged() {
+    this.emitStyle();
+  }
+
+  private emitStyle() {
+    const style: StyleEventDetail = {
       'interactive-disabled': this.disabled,
-    });
+    };
+
+    if (this.legacyFormController.hasLegacyControl()) {
+      style['radio-checked'] = this.checked;
+    }
+
+    this.ionStyle.emit(style);
   }
 
   private updateState = () => {
@@ -153,7 +211,96 @@ export class Radio implements ComponentInterface {
     this.ionBlur.emit();
   };
 
+  private get hasLabel() {
+    return this.el.textContent !== '';
+  }
+
+  private renderRadioControl() {
+    return (
+      <div class="radio-icon" part="container">
+        <div class="radio-inner" part="mark" />
+        <div class="radio-ripple"></div>
+      </div>
+    );
+  }
+
   render() {
+    const { legacyFormController } = this;
+
+    return legacyFormController.hasLegacyControl() ? this.renderLegacyRadio() : this.renderRadio();
+  }
+
+  private renderRadio() {
+    const { checked, disabled, inputId, color, el, justify, labelPlacement, inheritedAttributes, hasLabel } = this;
+    const mode = getIonMode(this);
+    const inItem = hostContext('ion-item', el);
+
+    return (
+      <Host
+        onClick={this.onClick}
+        class={createColorClasses(color, {
+          [mode]: true,
+          'in-item': inItem,
+          'radio-checked': checked,
+          'radio-disabled': disabled,
+          [`radio-justify-${justify}`]: true,
+          [`radio-label-placement-${labelPlacement}`]: true,
+          // Focus and active styling should not apply when the radio is in an item
+          'ion-activatable': !inItem,
+          'ion-focusable': !inItem,
+        })}
+      >
+        <label class="radio-wrapper">
+          {/*
+            The native control must be rendered
+            before the visible label text due to https://bugs.webkit.org/show_bug.cgi?id=251951
+          */}
+          <input
+            type="radio"
+            checked={checked}
+            disabled={disabled}
+            id={inputId}
+            ref={(nativeEl) => (this.nativeInput = nativeEl as HTMLInputElement)}
+            {...inheritedAttributes}
+          />
+          <div
+            class={{
+              'label-text-wrapper': true,
+              'label-text-wrapper-hidden': !hasLabel,
+            }}
+          >
+            <slot></slot>
+          </div>
+          <div class="native-wrapper">{this.renderRadioControl()}</div>
+        </label>
+      </Host>
+    );
+  }
+
+  private renderLegacyRadio() {
+    if (!this.hasLoggedDeprecationWarning) {
+      printIonWarning(
+        `ion-radio now requires providing a label with either the default slot or the "aria-label" attribute. To migrate, remove any usage of "ion-label" and pass the label text to either the component or the "aria-label" attribute.
+
+Example: <ion-radio>Option Label</ion-radio>
+Example with aria-label: <ion-radio aria-label="Option Label"></ion-radio>
+
+Developers can use the "legacy" property to continue using the legacy form markup. This property will be removed in an upcoming major release of Ionic where this form control will use the modern form markup.`,
+        this.el
+      );
+
+      if (this.legacy) {
+        printIonWarning(
+          `ion-radio is being used with the "legacy" property enabled which will forcibly enable the legacy form markup. This property will be removed in an upcoming major release of Ionic where this form control will use the modern form markup.
+
+Developers can dismiss this warning by removing their usage of the "legacy" property and using the new radio syntax.`,
+          this.el
+        );
+      }
+
+      this.hasLoggedDeprecationWarning = true;
+    }
+
     const { inputId, disabled, checked, color, el, buttonTabindex } = this;
     const mode = getIonMode(this);
     const { label, labelId, labelText } = getAriaLabel(el, inputId);
@@ -174,12 +321,10 @@ export class Radio implements ComponentInterface {
           interactive: true,
           'radio-checked': checked,
           'radio-disabled': disabled,
+          'legacy-radio': true,
         })}
       >
-        <div class="radio-icon" part="container">
-          <div class="radio-inner" part="mark" />
-          <div class="radio-ripple"></div>
-        </div>
+        {this.renderRadioControl()}
         <label htmlFor={inputId}>{labelText}</label>
         <input
           type="radio"