+++ core/src/components/segment/segment.tsx
@@ -3,11 +3,13 @@ import { Component, Element, Event, Host, Listen, Prop, State, Watch, h, writeTa
 
 import { config } from '../../global/config';
 import { getIonMode } from '../../global/ionic-global';
-import type { Color, SegmentChangeEventDetail, StyleEventDetail } from '../../interface';
+import type { Color, StyleEventDetail } from '../../interface';
 import type { Gesture, GestureDetail } from '../../utils/gesture';
 import { pointerCoord } from '../../utils/helpers';
 import { isRTL } from '../../utils/rtl';
 import { createColorClasses, hostContext } from '../../utils/theme';
+
+import type { SegmentChangeEventDetail } from './segment-interface';
 
 /**
  * @virtualProp {"ios" | "md"} mode - The mode determines which platform styles to use.
@@ -22,11 +24,9 @@ import { createColorClasses, hostContext } from '../../utils/theme';
 })
 export class Segment implements ComponentInterface {
   private gesture?: Gesture;
-  private didInit = false;
-  private checked?: HTMLIonSegmentButtonElement;
 
-  // Value to be emitted when gesture ends
-  private valueAfterGesture?: string;
+  // Value before the segment is dragged
+  private valueBeforeGesture?: string;
 
   @Element() el!: HTMLIonSegmentElement;
 
@@ -76,18 +76,15 @@ export class Segment implements ComponentInterface {
   /**
    * the value of the segment.
    */
-  @Prop({ mutable: true }) value?: string | null;
+  @Prop({ mutable: true }) value?: string;
 
   @Watch('value')
-  protected valueChanged(value: string | undefined, oldValue: string | undefined | null) {
+  protected valueChanged(value: string | undefined) {
+    /**
+     * `ionSelect` is emitted every time the value changes (internal or external changes).
+     * Used by `ion-segment-button` to determine if the button should be checked.
+     */
     this.ionSelect.emit({ value });
-    if (oldValue !== '' || this.didInit) {
-      if (!this.activated) {
-        this.ionChange.emit({ value });
-      } else {
-        this.valueAfterGesture = value;
-      }
-    }
 
     if (this.scrollable) {
       const buttons = this.getButtons();
@@ -128,7 +125,9 @@ export class Segment implements ComponentInterface {
   @Event() ionChange!: EventEmitter<SegmentChangeEventDetail>;
 
   /**
-   * Emitted when user has dragged over a new button
+   * Emitted when the value of the segment changes from user committed actions
+   * or from externally assigning a value.
+   *
    * @internal
    */
   @Event() ionSelect!: EventEmitter<SegmentChangeEventDetail>;
@@ -181,10 +180,10 @@ export class Segment implements ComponentInterface {
     if (this.disabled) {
       this.disabledChanged();
     }
-    this.didInit = true;
   }
 
   onStart(detail: GestureDetail) {
+    this.valueBeforeGesture = this.value;
     this.activate(detail);
   }
 
@@ -203,15 +202,32 @@ export class Segment implements ComponentInterface {
       this.addRipple(detail);
     }
 
-    const value = this.valueAfterGesture;
+    const value = this.value;
     if (value !== undefined) {
-      this.ionChange.emit({ value });
-      this.valueAfterGesture = undefined;
+      if (this.valueBeforeGesture !== value) {
+        this.emitValueChange();
+      }
     }
+    this.valueBeforeGesture = undefined;
+  }
+
+  /**
+   * Emits an `ionChange` event.
+   *
+   * This API should be called for user committed changes.
+   * This API should not be used for external value changes.
+   */
+  private emitValueChange() {
+    const { value } = this;
+    this.ionChange.emit({ value });
   }
 
   private getButtons() {
     return Array.from(this.el.querySelectorAll('ion-segment-button'));
+  }
+
+  private get checked() {
+    return this.getButtons().find((button) => button.value === this.value);
   }
 
   /**
@@ -327,9 +343,6 @@ export class Segment implements ComponentInterface {
     const buttons = this.getButtons();
     const index = buttons.findIndex((button) => button.value === this.value);
     const next = index + 1;
-
-    // Keep track of the currently checked button
-    this.checked = buttons.find((button) => button.value === this.value);
 
     for (const button of buttons) {
       button.classList.remove('segment-button-after-checked');
@@ -450,6 +463,10 @@ export class Segment implements ComponentInterface {
 
     this.value = current.value;
 
+    if (current !== previous) {
+      this.emitValueChange();
+    }
+
     if (this.scrollable || !this.swipeGesture) {
       if (previous) {
         this.checkButton(previous, current);
@@ -457,8 +474,6 @@ export class Segment implements ComponentInterface {
         this.setCheckedClasses();
       }
     }
-
-    this.checked = current;
   };
 
   private getSegmentButton = (selector: 'first' | 'last' | 'next' | 'previous'): HTMLIonSegmentButtonElement | null => {
@@ -515,8 +530,11 @@ export class Segment implements ComponentInterface {
     }
 
     if (keyDownSelectsButton) {
-      const previous = this.checked || current;
-      this.checkButton(previous, current);
+      const previous = this.checked;
+      this.checkButton(previous || current, current);
+      if (current !== previous) {
+        this.emitValueChange();
+      }
     }
     current.setFocus();
   }