+++ packages/vue-router/CHANGELOG.md
@@ -3,7 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.0.0](https://github.com/ionic-team/ionic/compare/v7.0.0-rc.5...v7.0.0) (2023-03-29)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
+# [7.0.0-rc.5](https://github.com/ionic-team/ionic/compare/v7.0.0-rc.4...v7.0.0-rc.5) (2023-03-29)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
 ## [6.7.1](https://github.com/ionic-team/ionic/compare/v6.7.0...v6.7.1) (2023-03-29)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
+# [7.0.0-rc.4](https://github.com/ionic-team/ionic/compare/v7.0.0-rc.3...v7.0.0-rc.4) (2023-03-27)
 
 **Note:** Version bump only for package @ionic/vue-router
 
@@ -19,6 +43,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-rc.3](https://github.com/ionic-team/ionic/compare/v7.0.0-rc.2...v7.0.0-rc.3) (2023-03-22)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
 ## [6.6.3](https://github.com/ionic-team/ionic/compare/v6.6.2...v6.6.3) (2023-03-22)
 
 **Note:** Version bump only for package @ionic/vue-router
@@ -27,7 +59,24 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-rc.2](https://github.com/ionic-team/ionic/compare/v7.0.0-rc.1...v7.0.0-rc.2) (2023-03-15)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
 ## [6.6.2](https://github.com/ionic-team/ionic/compare/v6.6.1...v6.6.2) (2023-03-15)
+
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
+# [7.0.0-rc.1](https://github.com/ionic-team/ionic/compare/v7.0.0-rc.0...v7.0.0-rc.1) (2023-03-08)
 
 **Note:** Version bump only for package @ionic/vue-router
 
@@ -43,6 +92,25 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-rc.0](https://github.com/ionic-team/ionic/compare/v7.0.0-beta.6...v7.0.0-rc.0) (2023-03-01)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
+# [7.0.0-beta.6](https://github.com/ionic-team/ionic/compare/v7.0.0-beta.5...v7.0.0-beta.6) (2023-03-01)
+
+
+### Bug Fixes
+
+* **angular:** devs can import ionic/angular styles in css ([#26875](https://github.com/ionic-team/ionic/issues/26875)) ([0d2fac2](https://github.com/ionic-team/ionic/commit/0d2fac2eefd610271fbbbecc36e844b6b289dfee))
+
+
+
+
+
 # [6.6.0](https://github.com/ionic-team/ionic/compare/v6.5.7...v6.6.0) (2023-03-01)
 
 **Note:** Version bump only for package @ionic/vue-router
@@ -52,6 +120,22 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 ## [6.5.7](https://github.com/ionic-team/ionic/compare/v6.5.6...v6.5.7) (2023-03-01)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
+# [7.0.0-beta.5](https://github.com/ionic-team/ionic/compare/v7.0.0-beta.4...v7.0.0-beta.5) (2023-02-27)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
+# [7.0.0-beta.4](https://github.com/ionic-team/ionic/compare/v7.0.0-beta.3...v7.0.0-beta.4) (2023-02-22)
 
 **Note:** Version bump only for package @ionic/vue-router
 
@@ -75,7 +159,23 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-beta.3](https://github.com/ionic-team/ionic/compare/v7.0.0-beta.2...v7.0.0-beta.3) (2023-02-15)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
 ## [6.5.4](https://github.com/ionic-team/ionic/compare/v6.5.3...v6.5.4) (2023-02-15)
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
+# [7.0.0-beta.2](https://github.com/ionic-team/ionic/compare/v7.0.0-beta.1...v7.0.0-beta.2) (2023-02-08)
 
 **Note:** Version bump only for package @ionic/vue-router
 
@@ -91,9 +191,33 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [7.0.0-beta.1](https://github.com/ionic-team/ionic/compare/v7.0.0-beta.0...v7.0.0-beta.1) (2023-02-01)
+
+
+**Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+
 ## [6.5.2](https://github.com/ionic-team/ionic/compare/v6.5.1...v6.5.2) (2023-02-01)
 
 **Note:** Version bump only for package @ionic/vue-router
+
+
+
+
+# [7.0.0-beta.0](https://github.com/ionic-team/ionic/compare/v6.5.1...v7.0.0-beta.0) (2023-01-25)
+
+
+### Code Refactoring
+
+* **vue:** only ship es modules ([#26054](https://github.com/ionic-team/ionic/issues/26054)) ([86bbed0](https://github.com/ionic-team/ionic/commit/86bbed07fc51fa1e3771f0198211c5064606e5bb)), closes [#25104](https://github.com/ionic-team/ionic/issues/25104)
+
+
+### BREAKING CHANGES
+
+* **vue:** `@ionic/vue` and `@ionic/vue-router` no longer ship a CommonJS entry point. Instead, only an ES Module entry point is provided for improved compatibility with Vite.